[lldb] Fix cast failure in FindFieldInStructArray
Complicated expressions in lldb broke the assumption that the expression is an identifier, thus we got a cast error.
This fix removes that from happening and also fixes the bug that if the identifier is an array itself the last index specified in the input is used to index that variable itself.

e.g 01 SAMPLE-TABLE.
           05 TABLE-DEPTH OCCURS 3 TIMES.
             10 TABLE-ROW OCCURS 3 TIMES.
               15 TABLE-COLUMN OCCURS 3 TIMES PIC 9(8).

Here TABLE-ROW(1, 2) means second element of TABLE-ROW OF TABLE-DEPTH(1).