[lldb] Fix bug in display of varying PLI strings
See 1156884604

The STG function also should include the prefix when counting the size, which for now is 2 bytes for all strings because the PLI compiler doesn't support COMPAT(V3) version.
If in the future we do support it, we would need to fix this again.

(cherry picked from commit 4b39f3e1b55c3df09f5cb89dcdd347682f790ba9)