[DWARF] Add support for RAINCODE vendor extension DW_OP_RC_resolve_file_address
This patch adds support for the DW_OP_RC_resolve_file_address opcode, a RAINCODE vendor
extension in DWARF. This opcode is responsible for resolving a file address to a load
address during expression evaluation. Specifically, it pops a file address from the DWARF
expression evaluation stack, resolves it to the corresponding load address, and pushes
the result back onto the stack.

This commit is added as a placeholder for maintaining backword compatibility with raincode
compiler, it will be removed in llvm20.