[LLDB] Patch 15: Add Level88 variable support
This patch adds support for evaluating Level-88 condition names in
COBOL expressions within LLDB.

COBOL can define a set of level-88 elements, which list potential values for
the preceding element. The level-88 statements can be used to test the value
of the preceding element.

For example, if the element FLAG1 has a value of ‘Y’, then FLAG1-ON resolves
to true when used to test the value of FLAG1.

01  BASIC-TEST-RECORD.
 02  FLAG1                  PIC X.
   88 FLAG1-ON              VALUE 'Y'.
   88 FLAG1-OFF             VALUE 'N'.